Add wrap_with_model_yaml decorator
This adds the decorator `wrap_with_model_yaml`, which takes
functions whose signature includes arguments typed as a pydantic model.
It adds to this function corresponding string arguments, presumed to point
towards yaml files which represent instances of the corresponding model.
These are then parsed as that model, and updated with any non-default arguments
passed by the model a la the original signature.
This amalgamated model is then passed to the original function.
This is meant to fill the same niche as the configargparse package, which provides
an analogous extension to argparse.